Refactor layoutmenu feature into separate zcml adn module
