+++ src/plone/app/mosaic/browser/configure.zcml
@@ -59,61 +59,15 @@
       template="templates/layouts-editor.pt"
       />
 
+  <!-- Include layoutmenu-feature and its ++layout++ traverser -->
+  <include
+      file="layoutsupport.zcml"
+      zcml:condition="have mosaic-layoutmenu"
+      />
   <adapter
       name="layout"
-      factory=".layoutsupport.DisplayLayoutTraverser"
+      factory=".layoutmenu.DisplayLayoutTraverser"
+      zcml:condition="not-have mosaic-layoutmenu"
       />
-
-  <configure zcml:condition="have mosaic-layout-menu">
-    <!-- Define the layout menu -->
-    <browser:menu
-        id="plone_contentmenu_layout"
-        title="The 'layout' menu - allows the user to select the predefined layout of an object"
-        class=".layoutsupport.DisplayLayoutMenu"
-        />
-
-    <!-- Hide the default -->
-    <adapter
-        for="plone.app.blocks.layoutbehavior.ILayoutAware
-             ..interfaces.IMosaicLayer"
-        name="plone.contentmenu.display"
-        factory=".layoutsupport.HiddenDisplaySubMenuItem"
-        provides="plone.app.contentmenu.interfaces.IContentMenuItem"
-        />
-
-    <!-- Display the custom menu -->
-    <adapter
-        for="plone.app.blocks.layoutbehavior.ILayoutAware
-             ..interfaces.IMosaicLayer"
-        name="plone.contentmenu.layout"
-        factory=".layoutsupport.DisplayLayoutSubMenuItem"
-        provides="plone.app.contentmenu.interfaces.IContentMenuItem"
-        />
-
-    <!-- Fix default view selection to allow ++layout++-traverser -->
-    <configure package="plone.app.content.browser"
-               zcml:condition="have plone-5">
-
-      <browser:page
-          for="plone.app.blocks.layoutbehavior.ILayoutAware"
-          layer="plone.app.mosaic.interfaces.IMosaicLayer"
-          name="select_default_view"
-          class="plone.app.mosaic.browser.selection.LayoutAwareDefaultViewSelectionView"
-          template="templates/select_default_view.pt"
-          permission="cmf.ModifyViewTemplate"
-          />
-
-      <browser:page
-          for="plone.app.blocks.layoutbehavior.ILayoutAware"
-          layer="plone.app.mosaic.interfaces.IMosaicLayer"
-          name="selectViewTemplate"
-          class="plone.app.mosaic.browser.selection.LayoutAwareDefaultViewSelectionView"
-          attribute="selectViewTemplate"
-          permission="cmf.ModifyViewTemplate"
-          />
-
-    </configure>
-
-  </configure>
 
 </configure>
